Support /etc/fstab mounting
`djmount' can't be used from /etc/fstab, because djmount considers any unknown options to be
an error, and exits.  `mount' supports a '-s' option for "sloppy" where unknown options are
ignored.

I've added both '-s' and '-o sloppy' to the latest (0.71-3) to engage behavior that reports,
but really ignores unknown options.  That is, this patch permits specification of 'sloppy'
as an /etc/fstab option, and thereafter permits unknown options to be logged as warnings, but
`djmount' continues.

Then, an /etc/fstab entry like:

djmount			/media/upnp		fuse	noauto,ro,user,allow_other,sloppy    0    0

succeeds, whereas prior, it would fail because 'standard' options like 'nodev' and 'nosuid' are
unknown to djmount as in:

[W] Found unknown option = noexec; ignoring
[W] Found unknown option = nosuid; ignoring
[W] Found unknown option = nodev; ignoring
[I]   Mount options = ro,noexec,nosuid,nodev,s,allow_other
[I] Charset : successfully initialised charset='UTF-8'
total 0

Link: https://sourceforge.net/tracker/?func=detail&aid=3514280&group_id=142039&atid=751317
Link: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=660098